refactor: introduce Custom Option
