+++ packages/main/src/Option.ts
@@ -1,12 +1,7 @@
-import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
-import slot from "@ui5/webcomponents-base/dist/decorators/slot.js";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 
-import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
-
-import optionCss from "./generated/themes/Option.css.js";
-import OptionTemplate from "./generated/templates/OptionTemplate.lit.js";
+import OptionBase from "./OptionBase.js";
 
 /**
  * @class
@@ -24,30 +19,8 @@ import OptionTemplate from "./generated/templates/OptionTemplate.lit.js";
  */
 @customElement({
 	tag: "ui5-option",
-	template: OptionTemplate,
-	renderer: litRender,
-	styles: [
-		optionCss,
-	],
 })
-class Option extends UI5Element {
-	/**
-	 * Defines the selected state of the component.
-	 * @default false
-	 * @public
-	 */
-	@property({ type: Boolean })
-	selected!: boolean;
-
-	/**
-	 * Defines the tooltip of the component.
-	 * @default ""
-	 * @private
-	 * @since 1.1.0
-	 */
-	@property()
-	title!: string;
-
+class Option extends OptionBase {
 	/**
 	 * Defines the `icon` source URI.
 	 *
@@ -61,15 +34,6 @@ class Option extends UI5Element {
 	icon?: string | null;
 
 	/**
-	 * Defines the value of the `ui5-select` inside an HTML Form element when this component is selected.
-	 * For more information on HTML Form support, see the `name` property of `ui5-select`.
-	 * @default ""
-	 * @public
-	 */
-	@property()
-	value!: string;
-
-	/**
 	 * Defines the additional text displayed at the end of the option element.
 	 * @default ""
 	 * @public
@@ -77,26 +41,6 @@ class Option extends UI5Element {
 	 */
 	@property()
 	additionalText!: string;
-
-	/**
-	 * Defines the focused state of the component.
-	 * @default false
-	 * @since 1.0.0-rc.13
-	 * @private
-	 */
-	@property({ type: Boolean })
-	focused!: boolean;
-
-	/**
-	 * Defines the content of the component.
-	 * @public
-	 */
-	@slot({ type: HTMLElement, "default": true })
-	content!: Array<HTMLElement>;
-
-	get stableDomRef() {
-		return this.getAttribute("stable-dom-ref") || `${this._id}-stable-dom-ref`;
-	}
 }
 
 Option.define();
