+++ packages/main/src/bundle.esm.ts
@@ -133,6 +133,8 @@ import ResponsivePopover from "./ResponsivePopover.js";
 import SegmentedButton from "./SegmentedButton.js";
 import SegmentedButtonItem from "./SegmentedButtonItem.js";
 import Select from "./Select.js";
+import Option from "./Option.js";
+import CustomOption from "./OptionCustom.js";
 import Slider from "./Slider.js";
 import SplitButton from "./SplitButton.js";
 import StepInput from "./StepInput.js";
