+++ packages/main/src/themes/OptionCustom.css
@@ -1,7 +1,6 @@
 :host {
 	height: var(--_ui5_list_item_base_height);
 	align-items: center;
-	height: 100%;
 	width: 100%;
 	padding: 0 1rem;
 }
