+++ packages/main/test/pages/SelectMenu.html
@@ -30,7 +30,7 @@
 						<ui5-avatar id="avatar" size="XS" initials="S"></ui5-avatar>
 					</div>
 
-					<ui5-option data-display-text="T-shirt [1]" data-additional-text="S" accessible-name="T-shirt size S" value="tshirt">
+					<ui5-option-custom data-display-text="T-shirt [1]" data-additional-text="S" accessible-name="T-shirt size S" value="tshirt">
 						<div class="selectMenuOption">
 							T-shirt [1]
 							<div>
@@ -38,9 +38,9 @@
 								<ui5-icon name="employee"></ui5-icon>
 							</div>
 						</div>
-					</ui5-option>
+					</ui5-option-custom>
 			
-					<ui5-option data-display-text="Dress [2]" value="dress" data-additional-text="M" accessible-name="Dress size M" style="background-color: red; color:yellow;">
+					<ui5-option-custom data-display-text="Dress [2]" value="dress" data-additional-text="M" accessible-name="Dress size M" style="background-color: red; color:yellow;">
 						<div class="selectMenuOption selectMenuOption--warning">
 							Dress [2]
 							<div>
@@ -48,9 +48,9 @@
 								<ui5-icon name="employee" class="selectMenuOptionIcon--warning"></ui5-icon>
 							</div>
 						</div>
-					</ui5-option>
+					</ui5-option-custom>
 			
-					<ui5-option data-display-text="Skirt [3]" value="skirt" data-additional-text="L" accessible-name="Skirt size L" selected>
+					<ui5-option-custom data-display-text="Skirt [3]" value="skirt" data-additional-text="L" accessible-name="Skirt size L" selected>
 						<div  class="selectMenuOption">
 							Skirt [3]
 							<div>
@@ -58,7 +58,7 @@
 								<ui5-icon name="employee"></ui5-icon>
 							</div>
 						</div>
-					</ui5-option>
+					</ui5-option-custom>
 
 				</ui5-select>
 	
@@ -627,8 +627,8 @@
 				</ui5-select>
 			
 				<ui5-select id="errorSelect" value-state="Error">
-					<ui5-option selected>Cozy</ui5-option>
-					<ui5-option selected>Compact</ui5-option>
+					<ui5-option>Cozy</ui5-option>
+					<ui5-option>Compact</ui5-option>
 					<ui5-option selected>Condensed</ui5-option>
 				</ui5-select>
 			</section>
